refactor(#70): 서버 분리를 위한 메서드 구조 변경
스케줄러 서버에서 알림 서버로 두번 요청하는 구조->메서드 하나로 통합 후 한 번 요청으로 변경
